requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
celery-types==0.23.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2025.6
distlib==0.3.8
Django==5.2.1
django-crispy-forms==2.4
docker==7.1.0
filelock==3.15.4
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
Markdown==3.8
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.26.6
wcwidth==0.2.13
zstandard==0.22.0
